+++ options.go
@@ -32,3 +32,12 @@ func WithMessageParams(params slack.PostMessageParameters) Option {
 		return nil
 	}
 }
+
+// WithLogUnknownMessageTypes makes the adapter log unknown message types as
+// error message for debugging. This option is disabled by default.
+func WithLogUnknownMessageTypes() Option {
+	return func(conf *Config) error {
+		conf.LogUnknownMessageTypes = true
+		return nil
+	}
+}