+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-Nothing so far
+- Update Slack client library to fix RTM unmarshal errors
+- Add new `WithLogUnknownMessageTypes` option to help debug issues with Slack
+- Log slack errors when RTM message parsing fails
 
 ## [v0.7.0] - 2019-10-22
 - Support sending and receiving reactions